% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust.R
\name{adjust.linearly}
\alias{adjust.linearly}
\title{ADJUST LINEARLY}
\usage{
adjust.linearly(formula, data)
}
\arguments{
\item{formula}{A formula that gives the relationship of the adjustment.}

\item{data}{A dataframe which contains the dependent and the independent variables.}
}
\value{
A dataframe containing a new column with the adjusted values.
}
\description{
Adjust a dependent variable y linearly to several dependent variables x1, x2 ...
}
\examples{

hlpr4life::load.pkgs(c("ggplot2","reshape2"))
mtcrs<-mtcars
mtcrs$am<-factor(mtcrs$am,labels=c("auto","manu"))
mtcrs<-adjust.linearly(disp~mpg*am,mtcrs)
mtcrs$id<-rownames(mtcrs)
mtcrs.mlt<-melt(mtcrs[,c("mpg","cyl","disp","disp.adj.for.mpg.am","am","id")],c("mpg","cyl","am","id"))
mtcrs.mlt<-hlpr4life::rename.columns(mtcrs.mlt,c("value","am"),c("displacement","gearbox"))
ggplot(mtcrs.mlt,aes(mpg,displacement,col=gearbox,label=paste(id,c("orig","adj")[match(variable,c("disp","disp.adj.for.mpg.am"))]),group=id),mtcrs)+
theme_bw()+geom_line()+geom_text(aes(col=paste(gearbox,variable)))+theme(legend.position="none")
}
