% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.colors.R
\name{color.gradient}
\alias{color.gradient}
\title{COLOR GRADIENT}
\usage{
color.gradient(col.from = "red", col.to = "green", length.out = 2)
}
\arguments{
\item{col.from}{first color in sequence}

\item{col.to}{last color in sequence}

\item{length.out}{length of sequence}
}
\value{
dataframe with color vector, a color sequence from col.from to col.to and sequence from 0 to 1
}
\description{
calculates a data frame containing a sequence of colors from one color to a second one and a parallel sequence from 0 to 1
}
\examples{
color.gradient( "red", "green", 16 )
}
