% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piano.R
\name{freq.of.key}
\alias{freq.of.key}
\title{freq.of.key}
\usage{
freq.of.key(key = 25)
}
\arguments{
\item{key}{key for what the frequency is asked for,
key number of a piano,
most left is 1}
}
\value{
frequency of the key
}
\description{
freq.of.key returns the frequency that a piano key plays which first (most left) key plays tone "A0"
}
\examples{
freq.of.key( 25 )
}
