% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piano.R
\name{freq.of.note}
\alias{freq.of.note}
\title{freq.of.note}
\usage{
freq.of.note(note)
}
\arguments{
\item{note}{note for what the frequency is asked for,}
}
\value{
frequency of note
}
\description{
freq.of.note gives the frequency for a certain note
}
\examples{
freq.of.note( "A2" )
}
