% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/life.tools.R
\name{get.columns}
\alias{get.columns}
\title{GET COLUMNS}
\usage{
get.columns(data, pattern = "DAT|SIC|GROUP", perl = T)
}
\arguments{
\item{data}{dataframe which has some date columns}

\item{pattern}{search pattern for finding column names via grep}
}
\value{
names of searched columns
}
\description{
searches for column names that matches a given the pattern.
}
\examples{
(d<-data.frame(SIC="LI01234",Y=1000.,x=10,SCI_GROUP="A2_12",DATE="2017-10-05",EDAT="2017-10-04"))
get.columns(d)
}
