% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/life.tools.R
\name{get.info}
\alias{get.info}
\title{GET INFO FOR DATAFRAME}
\usage{
get.info(data)
}
\arguments{
\item{data}{dataframe which colums should be lightly summarised}
}
\value{
dataframe with the total number, the numbers of available and missing data for each column
}
\description{
GET INFO FOR DATAFRAME
}
\examples{
(d<-data.frame(x=c(NA,"Hello",NA,"World",NA),y=c(1:5),z=rep(NA,5)))
get.info( d )
}
