% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/life.tools.R
\name{get.merging.infos}
\alias{get.merging.infos}
\title{GET MERGING INFOS}
\usage{
get.merging.infos(table.names = ls())
}
\arguments{
\item{table.names}{}
}
\value{
data frame with merging informations for all given tables
}
\description{
GET MERGING INFOS
}
\examples{
(a<-data.frame(SGROUP="A2_02",DATE="200-10-05",Sic="LI12345678"))
(b<-data.frame(GRUPPE="A3_02",DATUM="200-10-05",PSEUDONYM="LI12345678"))
(c<-data.frame(GRP="A2_02",DATE="200-10-05",EDAT="200-10-04",PSEUDO="LI12345679",edat.new="2017.10-03"))
(infos<-get.merging.infos(c("a","b","c")))
infos$SIC
infos$SCI_GROUP
infos$DATE
}
