% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/life.tools.R
\name{get.scigroup.columns}
\alias{get.scigroup.columns}
\title{GET SCI-GROUP COLUMNS}
\usage{
get.scigroup.columns(data,
  pattern = "sci_group|sci-group|scigroup|sgroup|group|grp|gruppe",
  perl = T)
}
\arguments{
\item{data}{dataframe which has some sci-group columns}

\item{perl}{logical: use perl regex}
}
\value{
names of sci-group columns
}
\description{
searches for some sci-group like column names. one has the opportunity to give a
search string for a certain pattern.
}
\examples{
(d<-data.frame(SGROUP="A2_02",SCI_GROUP="B1_10",Gruppe="A1-SK_10",GRP="A3_09"))
get.scigroup.columns(d)
}
