% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/life.tools.R
\name{get.sic.columns}
\alias{get.sic.columns}
\title{GET SIC COLUMNS}
\usage{
get.sic.columns(data, pattern = "sic|pseudo")
}
\arguments{
\item{data}{dataframe which has some sic columns}

\item{pattern}{search pattern for finding column names via grep}
}
\value{
names of sic columns
}
\description{
GET SIC COLUMNS
}
\examples{
(d<-data.frame(SIC="LI12345678",sic="LI12345679",PSEUDO="LI1234567X",PSEUDONYM="LI12345670"))
get.sic.columns(d)
}
