% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsubplot.R
\name{ggsubplot}
\alias{ggsubplot}
\title{ggsubplot}
\usage{
ggsubplot(..., plotlist = NULL, cols = 1, layout = NULL)
}
\arguments{
\item{...}{end of list of ggplots}

\item{plotlist}{end of list of ggplots}

\item{cols}{count of columns}

\item{layout}{matrix for the layout with indices of the plots}
}
\value{
subplots
}
\description{
gives the opportunity for subplots
}
