% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/life.tools.R
\name{key.col}
\alias{key.col}
\title{KEY.COLS}
\usage{
key.col(col.x, col.y, sep = "~")
}
\arguments{
\item{col.x}{lhs column name}

\item{col.y}{rhs column name}

\item{sep}{a separator for binding column contents}
}
\value{
a key combined out of given columns
}
\description{
key.cols creates a key out of several columns
}
\examples{
key.col( letters[3:1], LETTERS[1:3],"~")
}
