% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/life.tools.R
\name{key.df}
\alias{key.df}
\title{KEY.DF}
\usage{
key.df(data, column.names, sep = "~")
}
\arguments{
\item{data}{a dataframe}

\item{column.names}{column names that should be used for creating a unique key}

\item{sep}{a separator for binding column contents}
}
\value{
a key combined out of given columns
}
\description{
key.df creates a key out of several columns of a dataframe
}
\examples{
key.df(data.frame(x=letters[ runif(10,1,10)],y=LETTERS[runif(10,1,10)],z=rnorm(10)),c("x","y"),"~")
}
