% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piano.R
\name{key.of.freq}
\alias{key.of.freq}
\title{KEY OF FREQUENCY}
\usage{
key.of.freq(frq = 110)
}
\arguments{
\item{frq}{frequency of the key}
}
\value{
key of frq
}
\description{
key.of.freq returns the piano key that plays a tone of the frequency frq when its first (most left) key plays tone "A0"
}
\examples{
key.of.freq( 110 )
}
