% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piano.R
\name{key.of.note}
\alias{key.of.note}
\title{KEY OF NOTE}
\usage{
key.of.note(note)
}
\arguments{
\item{note}{note for what the key is asked for}
}
\value{
key of note
}
\description{
key.of.note gives the note for a certain piano key which first (most left) key plays tone "A0"
}
\examples{
key.of.note( c( "C0", "D0", "E0", "F0", "G0", "A1", "B1", "C1" ) )
}
