% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/life.tools.R
\name{list.append}
\alias{list.append}
\title{LIST APPEND}
\usage{
list.append(list, x, name = NA)
}
\arguments{
\item{x}{element that should be append to list lst}

\item{name}{optinal a name for list element}

\item{lst}{list that should extended by the element x}
}
\value{
by x extended list
}
\description{
list.append gives the opportunity to append elements to a list and name it.
}
\examples{
( lst <- list( x = 9 ) )
( lst <- list.append( lst, value = Sys.time( ), name = "TIME" ) )
( lst <- list.append( lst, "y", "Ypsilon" ) )
( lst <- list.append( lst, "unnamed" ) )
( lst <- list.append( lst, c( "A", "B" ), c( "a", "b" ) ) )
}
