% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/life.tools.R
\name{load.pkgs}
\alias{load.pkgs}
\title{LOAD PACKAGES}
\usage{
load.pkgs(pkgs = c("hlpr4life"))
}
\arguments{
\item{pkgs}{}
}
\description{
load.pkgs tries to load a package into the environment.
If this fails, it tries to install the package from CRAN-mirror and loads it afterwards.
}
\examples{
hlpr4life::load.pkgs(c("hlpr4life","dplyr","ggplot2","ggthemes","reshape2"))
}
