% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/life.tools.R
\name{make.sics}
\alias{make.sics}
\title{make.sics}
\usage{
make.sics(n.size = 100, n.first = 1, n.last = n.size)
}
\arguments{
\item{n}{count of sics that shall be created}
}
\value{
vector of characters of sics
}
\description{
make.sics
}
\examples{
make.sics( 90, 110 )
}
