% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.likely.R
\name{merge.likely}
\alias{merge.likely}
\title{merge likely}
\usage{
\method{merge}{likely}(d1, d2, by = NULL, by.x = NULL, by.y = NULL,
  by.lk = NULL, by.lk.x = NULL, by.lk.y = NULL, min = NULL,
  max = NULL, all = F, all.x = F, all.y = F, reorder.names = T,
  trim = F, add.diffs = F)
}
\arguments{
\item{d1}{left hand side dataframe}

\item{d2}{right hand side dataframe}

\item{by}{vector of column names for exact merging via equal column names}

\item{by.x}{vector of column names for the left hand side dataframe for exact merging via different column names}

\item{by.y}{vector of column names for the right hand side dataframe for exact merging via different column names}

\item{by.lk}{vector of column names for likely merging via equal column names}

\item{by.lk.x}{vector of column names for the left hand side dataframe for likely merging via different column names}

\item{by.lk.y}{vector of column names for the right hand side dataframe for likely merging via different column names}

\item{min}{vector of minimum limits for likely merging}

\item{max}{vector of maximum limits for likely merging}

\item{all}{merge and add remaining columns from both dataframes, default is FALSE}

\item{all.x}{merge and add remaining columns from left dataframes, default is FALSE}

\item{all.y}{merge and add remaining columns from right dataframes, default is FALSE}

\item{reorder.names}{gives the opportunity to reorder columns that way the pairs participated on merging lay next to each other. Doen't use it with trim = T! It makes no sence., default is TRUE}

\item{trim}{if trim is TRUE it selects the left hand side column for final dataframe and deletes the right hand side columns, default is FALSE}

\item{add.diffs}{if add.diffs is TRUE differences of likely merged columns are added to resulting dataframe, default is FALSE}
}
\value{
merged data frame
}
\description{
merge likely merges two date frames by exact matching also by nearly matching. If the difference of two variable A of dataframe 1 and variable B of dataframe 2 may lay in an interval, one has the opportunity to give the limits of the interval by min and max.
}
\examples{
d1<-data.frame(w=sample(1:10,20,T),x=sample(1:10,20,T),y=sample(1:10,20,T),z=sample(c(1:10),20,T),id=1)
d2<-data.frame(w=sample(1:10,20,T),x=sample(1:10,20,T),Y=sample(1:10,20,T),C=sample(c(1:10),20,T),id=2)
merge.likely(d1=d1,d2=d2,by=c("x"),by.lk=c("w","id"),by.lk.x=c("y"),by.lk.y=c("Y"),min=c(-2,-2,-2),max=c(1,1,1),reorder.names=F,trim=F,add.diffs=F)
merge.likely(d1=d1,d2=d2,by=c("x"),by.lk=c("w","id"),by.lk.x=c("y"),by.lk.y=c("Y"),min=c(-2,-2,-2),max=c(1,1,1),reorder.names=T,trim=F,add.diffs=F)
merge.likely(d1=d1,d2=d2,by=c("x"),by.lk=c("w","id"),by.lk.x=c("y"),by.lk.y=c("Y"),min=c(-2,-2,-2),max=c(1,1,1),reorder.names=F,trim=T,add.diffs=F)
merge.likely(d1=d1,d2=d2,by=c("x"),by.lk=c("w","id"),by.lk.x=c("y"),by.lk.y=c("Y"),min=c(-2,-2,-2),max=c(1,1,1),reorder.names=T,trim=T,add.diffs=F)
merge.likely(d1=d1,d2=d2,by=c("x"),by.lk=c("w","id"),by.lk.x=c("y"),by.lk.y=c("Y"),min=c(-2,-2,-2),max=c(1,1,1),reorder.names=F,trim=F,add.diffs=T)
merge.likely(d1=d1,d2=d2,by=c("x"),by.lk=c("w","id"),by.lk.x=c("y"),by.lk.y=c("Y"),min=c(-2,-2,-2),max=c(1,1,1),reorder.names=T,trim=F,add.diffs=T)
merge.likely(d1=d1,d2=d2,by=c("x"),by.lk=c("w","id"),by.lk.x=c("y"),by.lk.y=c("Y"),min=c(-2,-2,-2),max=c(1,1,1),reorder.names=F,trim=T,add.diffs=T)
merge.likely(d1=d1,d2=d2,by=c("x"),by.lk=c("w","id"),by.lk.x=c("y"),by.lk.y=c("Y"),min=c(-2,-2,-2),max=c(1,1,1),reorder.names=T,trim=T,add.diffs=T)
}
