% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.likely.R
\name{ml}
\alias{ml}
\title{merge likely}
\usage{
ml(d1, d2, by = NULL, by.x = NULL, by.y = NULL, by.lk = NULL,
  by.x.lk = NULL, by.y.lk = NULL, min = NULL, max = NULL,
  reorder.names = T, trim = F, add.diffs = F)
}
\arguments{
\item{d1}{left hand side dataframe}

\item{d2}{right hand side dataframe}

\item{by}{vector of column names for exact merging via equal column names}

\item{by.x}{vector of column names for the left hand side dataframe for exact merging via different column names}

\item{by.y}{vector of column names for the right hand side dataframe for exact merging via different column names}

\item{by.lk}{vector of column names for likely merging via equal column names}

\item{by.x.lk}{vector of column names for the left hand side dataframe for likely merging via different column names}

\item{by.y.lk}{vector of column names for the right hand side dataframe for likely merging via different column names}

\item{min}{vector of minimum limits for likely merging}

\item{max}{vector of maximum limits for likely merging}

\item{reorder.names}{gives the opportunity to reorder columns that way the pairs participated on merging lay next to each other. Doen't use it with trim = T! It makes no sence.}

\item{trim}{if trim is TRUE it selects the left hand side column for final dataframe and deletes the right hand side columns}
}
\value{
merged data frame
}
\description{
merge likely merges two date frames by exact matching also by nearly matching. If the difference of two variable A of dataframe 1 and variable B of dataframe 2 may lay in an interval, one has the opportunity to give the limits of the interval by min and max.
}
\examples{
d1<-data.frame(w=sample(1:10,20,T),x=sample(1:10,20,T),y=sample(1:10,20,T),z=sample(c(1:10),20,T),id=1)
d2<-data.frame(w=sample(1:10,20,T),x=sample(1:10,20,T),Y=sample(1:10,20,T),C=sample(c(1:10),20,T),id=2)
ml(d1=d1,d2=d2,by=c("x"),by.lk=c("w","id"),by.x.lk=c("y"),by.y.lk=c("Y"),min = c(-2,-2,-2),max=c(1,1,1),reorder.names = F,trim = F,add.diffs=F)
ml(d1=d1,d2=d2,by=c("x"),by.lk=c("w","id"),by.x.lk=c("y"),by.y.lk=c("Y"),min = c(-2,-2,-2),max=c(1,1,1),reorder.names = T,trim = F,add.diffs=F)
ml(d1=d1,d2=d2,by=c("x"),by.lk=c("w","id"),by.x.lk=c("y"),by.y.lk=c("Y"),min = c(-2,-2,-2),max=c(1,1,1),reorder.names = F,trim = T,add.diffs=F)
ml(d1=d1,d2=d2,by=c("x"),by.lk=c("w","id"),by.x.lk=c("y"),by.y.lk=c("Y"),min = c(-2,-2,-2),max=c(1,1,1),reorder.names = T,trim = T,add.diffs=F)
ml(d1=d1,d2=d2,by=c("x"),by.lk=c("w","id"),by.x.lk=c("y"),by.y.lk=c("Y"),min = c(-2,-2,-2),max=c(1,1,1),reorder.names = F,trim = F,add.diffs=T)
ml(d1=d1,d2=d2,by=c("x"),by.lk=c("w","id"),by.x.lk=c("y"),by.y.lk=c("Y"),min = c(-2,-2,-2),max=c(1,1,1),reorder.names = T,trim = F,add.diffs=T)
ml(d1=d1,d2=d2,by=c("x"),by.lk=c("w","id"),by.x.lk=c("y"),by.y.lk=c("Y"),min = c(-2,-2,-2),max=c(1,1,1),reorder.names = F,trim = T,add.diffs=T)
ml(d1=d1,d2=d2,by=c("x"),by.lk=c("w","id"),by.x.lk=c("y"),by.y.lk=c("Y"),min = c(-2,-2,-2),max=c(1,1,1),reorder.names = T,trim = T,add.diffs=T)
}
