% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piano.R
\name{note.of.freq}
\alias{note.of.freq}
\title{note.of.freq}
\usage{
note.of.freq(frq = 110)
}
\arguments{
\item{frq}{frequency for what the note is asked for}
}
\value{
note of frequency
}
\description{
note.of.freq gives the note for a certain frequency
}
\examples{
note.of.freq( 110 )
}
