% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piano.R
\name{note.of.key}
\alias{note.of.key}
\title{NOTE OF KEY}
\usage{
note.of.key(key = c(4, 6, 8, 9, 11, 13, 15, 16))
}
\arguments{
\item{key}{key for what the note is asked for,}
}
\value{
note of key
}
\description{
note.of.key gives the note for a certain piano key which first (most left) key plays tone "A0"
}
\examples{
note.of.key( 25 )
}
