% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piano.R
\name{piano}
\alias{piano}
\title{PIANO}
\usage{
piano(left.key = 4, right.key = 88)
}
\arguments{
\item{left.key}{left key of piano}

\item{right.key}{right key of piano}
}
\value{
piano with keys, their colors, notes and frequencies
}
\description{
piano gives a data.frame of a piano which first (most left) key plays tone "A0".
One can expand or shrink the keyboard by giving the piano's left.key, right.key.
The returned dataframe then contains all keys between left.key and right.key, their colors, notes and played frequencies.
}
\examples{
load.pkgs( c( "ggplot2", "ggthemes" ) )
piano( 4, 4 + 2 * 12 )
ggplot( piano( 4, 4 + 2 * 12 ) ) +
    geom_histogram( aes( note, -c( 1, .63 )[ match( color, c( "ivory", "ebony" ) ) ], fill = color ), stat = "identity" ) +
    scale_fill_manual( values = c( "#000000", "#f0f3f4" ), guide = F ) +
    geom_text( aes( note, label = note, col = color ), y = -.2, angle = 90 ) +
    scale_color_manual( values = c( "#f0f3f4", "#000000" ), guide = F ) +
    theme_solid( fill = "#405060" ) +
    theme(
        axis.text.x = element_blank( ),
        axis.title.x = element_blank( ),
        axis.text.y = element_blank( ),
        axis.title.y = element_blank( ) ) +
    ylim( -1, 0 )
}
