% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{plot.color.gradient}
\alias{plot.color.gradient}
\title{PLOT COLOR GRADIENT}
\usage{
\method{plot}{color.gradient}(col.from = "red", col.to = "yellow",
  length.out = 8, dir = "vertical", verbose = T)
}
\arguments{
\item{col.from}{first color}

\item{col.to}{last color}

\item{length.out}{number of colors}

\item{dir}{direction can be "horizontal" or "vertical"}

\item{verbose}{if true colors are printed inside color rectangles}
}
\value{
a ggplot
}
\description{
plots a gradient picture
}
\examples{
ggsubplot(plot.color.gradient(),plot.color.gradient(dir="h"),cols=2)
}
