% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.colors.R
\name{plot.colors}
\alias{plot.colors}
\title{PLOT COLORS}
\usage{
\method{plot}{colors}()
}
\value{
a plot of all colors returned by colors( )
}
\description{
a plot of all colors returned by colors( )
}
\examples{
plot.colors( )
}
