% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyfy.R
\name{polyfy}
\alias{polyfy}
\title{polyfy}
\usage{
polyfy(x, y, g)
}
\arguments{
\item{x}{vector of numbers or factors}

\item{y}{values of numbers}

\item{g}{grouping values}
}
\value{
data frame with rows X, Y, G
}
\description{
polyfy
}
\examples{
polyfy( c( 1 : 100 ), rnorm( 100 ), sample( c( "A", "B" ), 100, TRUE ) )
}
