% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/life.tools.R
\name{remove.columns}
\alias{remove.columns}
\title{REMOVE COLUMNS}
\usage{
remove.columns(data, column.names)
}
\arguments{
\item{data}{A dataframe that contains columns which should be removed.}

\item{column.names}{Names of columns that should be removed.}
}
\value{
A dataframe without removed columns.
}
\description{
removes columns out of a dataframe
}
\examples{
remove.columns(data.frame(x="X",y="Y",z="Z",w="W"),c("x","z"))
}
