% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/life.tools.R
\name{remove.list.elements}
\alias{remove.list.elements}
\title{REMOVE LIST ELEMENTS}
\usage{
remove.list.elements(list, names)
}
\arguments{
\item{names}{Names of elements that should be removed.}

\item{data}{A list that contains named elements which should be removed.}
}
\value{
A list without removed columns.
}
\description{
removes elements out of a list
}
\examples{
remove.list.elements(list(x="X",y="Y",z="Z",w="W"),c("x","z"))
}
