% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename.R
\name{rename.column}
\alias{rename.column}
\title{RENAME COLUMN}
\usage{
rename.column(data, old.column.name, new.column.name)
}
\arguments{
\item{data}{dataframe for which one name should be changed}

\item{old.column.name}{old column name}

\item{new.column.name}{new column name}
}
\value{
dataframe with new column name
}
\description{
rename.column( data, "old", "new" ) changes the old column name "old" to the new one "new".
}
\examples{
(d<-rename.column(data.frame(x=c(1:10),y=rnorm(10)),"x","X"))
(d<-rename.column(d,"y","Y"))
}
