% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/life.tools.R
\name{rename.list.elements}
\alias{rename.list.elements}
\title{RENAME LIST}
\usage{
\method{rename}{list.elements}(list, old.names, new.names)
}
\arguments{
\item{list}{named list}

\item{old.names}{old names}

\item{new.names}{new names}
}
\value{
dataframe with replaced column names
}
\description{
rename.list.elements takes a list containing the old element names "alter" and "groesse" and returns a new list with the replaced names "age" and "size".
}
\details{
RENAME LIST ELEMENTS
}
\examples{
(l<-rename.list.elements(list(x="x",y="y",z="z"),c("y","z","x"),c("Ypsilon","CED","U")))
}
