% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/life.tools.R
\name{some.sics}
\alias{some.sics}
\title{some.sics}
\usage{
some.sics(n.size = 100, n.first = 1, n.last = n.size,
  digits = ceiling(log10(n.last)))
}
\arguments{
\item{n.size}{count of sics that shall be created}

\item{n.first}{first sic to be created}

\item{n.last}{last sic to be created}
}
\value{
character vector of sics
}
\description{
some.sics creates a vector of coherently sics of same width.
}
\examples{
some.sics( n.first = 0, n.last = 110 )
some.sics( n.size = 12 )
some.sics( 10 )

}
