% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/life.tools.R
\name{sum.av}
\alias{sum.av}
\title{SUM AVAILABLES}
\usage{
\method{sum}{av}(data)
}
\arguments{
\item{data}{dataframe for which availables in columns should be summed up}
}
\value{
dataframe with the number of available data for each column
}
\description{
gives the number of the available values of every column of a dataframe
}
\examples{
(d<-data.frame(x=c(NA,"Hello",NA,"World",NA),y=c(1:5),z=rep(NA,5)))
sum.av(d)
}
