% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/life.tools.R
\name{table.columns}
\alias{table.columns}
\title{TABLE COLUMNS}
\usage{
table.columns(data, show.missings = F)
}
\arguments{
\item{data}{dataframe which columns should be summarised}

\item{show.missings}{flag for summing missings or availables}
}
\value{
print of missings or availables, depending on show.missings
}
\description{
TABLE COLUMNS
}
\examples{
table.columns(data.frame(x=c(1:3),y=c(NA,1,NA),z=c(NA,NA,NA)))
table.columns(data.frame(x=c(1:3),y=c(NA,1,NA),z=c(NA,NA,NA)),F)
}
