% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/life.tools.R
\name{table.data}
\alias{table.data}
\title{TABLE DATA}
\usage{
table.data(data)
}
\arguments{
\item{data}{dataframe which columns should be summarised}
}
\value{
summary of data
}
\description{
TABLE DATA
}
\examples{
table.data(data.frame(x=c(1:3),y=c(NA,1,NA),z=c(NA,NA,NA)))
table.data(data.frame(x=c(1:3),y=c(NA,1,NA),z=c(NA,NA,NA)))
}
