% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/life.tools.R
\name{table.df}
\alias{table.df}
\title{TABLE DATAFRAME}
\usage{
table.df(data, summary = F, as.dataframe = !summary)
}
\arguments{
\item{data}{dataframe which columns should be summarised}

\item{summary}{logical: if is true additionally min, median, mean and max will be tabled}

\item{as.data.frame}{logical: result is shown as data frame}
}
\value{
table of data containing sum of missing and available data their mins and maxs.
}
\description{
TABLE DATAFRAME
}
\examples{
table.df(data.frame(x=c(1:3),y=c(NA,1,NA),z=c(NA,NA,NA)))
}
