% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/life.tools.R
\name{calendar}
\alias{calendar}
\title{calendar}
\usage{
calendar(start = "2000-01-01", end = "2017-06-19", abbreviate = T,
  tz = "Europe/Berlin")
}
\arguments{
\item{start}{first date}

\item{end}{last date}

\item{abbreviate}{if true the names of weekdays will be abbreviated}
}
\value{
a calendar from start date to end date
}
\description{
calendar
}
\examples{
calendar( )
}
