% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsubplot.R
\name{ggsubplot}
\alias{ggsubplot}
\title{ggsubplot}
\usage{
ggsubplot(..., plotlist = NULL, cols = 1, layout = NULL)
}
\arguments{
\item{...}{list of ggplots}

\item{plotlist}{last plots}

\item{cols}{number of columns}

\item{layout}{layout ist a matrix containing the index of a certain ggplot}
}
\value{
a plotmatrix containing all ggplots
}
\description{
ggsubplot
}
\examples{
library( ggplot2 )
ggsubplot( ggplot(), ggplot(), cols = 1 )
}
