% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsubplot.R
\name{ggsubplot}
\alias{ggsubplot}
\title{GGSUBPLOT}
\usage{
ggsubplot(..., plotlist = NULL, cols = 1, layout = NULL)
}
\arguments{
\item{...}{end of list of ggplots}

\item{plotlist}{end of list of ggplots}

\item{cols}{count of columns}

\item{layout}{matrix for the layout with indices of the plots}
}
\value{
subplots
}
\description{
gives the opportunity for subplots. it has nothing to do with ggplot exept of the fact, that it is build for ggplots. It is taken from R-Cookbook.
}
\examples{
hlpr4life::load.pkgs(c("hlpr4life","ggplot2","dplyr"))
(d<-data.frame(x=rnorm(1000,10),y=sample(letters[1:3],1000,T),s=sample(c("female","male"),1000,T)))
m<-summarise(group_by(d,y,s),m=mean(x))
ggsubplot(
  ggplot( d ) + theme_bw( ) + scale_color_discrete( guide = F ) +
    geom_point( aes( c( 1 : 1000 ), x, col = paste0( y, s ) ) ),
  ggplot( d ) +
    theme_bw( ) + scale_fill_discrete( guide = F ) + scale_color_discrete( guide = F ) +
    geom_histogram( aes( x, fill = paste0( y, s ) ), alpha = .5 ) +
    geom_vline( aes( xintercept = m, col = paste0( y, s ) ), m ) +
    facet_grid( s ~ y ),
  ggplot( d ) +
    theme_bw( ) + scale_fill_discrete( guide = F ) + scale_color_discrete( guide = F ) +
    geom_histogram( aes( x, fill = paste0( y, s ) ), alpha = .5 ) +
    geom_vline( aes( xintercept = m, col = paste0( y, s ) ), m ),
  layout = t(
    matrix(
      c(
        1, 2,
        3, 3 ),
      ncol = 2 ) ) )
}
