% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.pkgs.R
\name{load.pkgs}
\alias{load.pkgs}
\title{load.pkgs}
\usage{
load.pkgs(pkgs)
}
\arguments{
\item{pkgs}{}
}
\description{
load packages tries to load a package into the environment. If it fails, it tries to install the package and loads it afterwards.
}
\examples{
load.pkgs( c( "dplyr", "ggplot2", "ggthemes", "reshape2" ) )
}
