% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/life.tools.R
\name{rename.columns}
\alias{rename.columns}
\title{RENAME COLUMNS}
\usage{
\method{rename}{columns}(data, old.column.names, new.column.names)
}
\arguments{
\item{data}{dataframe with new column names}

\item{old.column.names}{old column names}

\item{new.column.names}{new column names}
}
\value{
dataframe with replaced column names
}
\description{
rename.columns( data, c( "alter", "groesse" ), c( "age", "height" ) ) takes a data frame d containing the old column name "alter" and "groesse" and returns a new dataframe with the replaced names "age" and "size".
}
\examples{
(d<-rename.columns(d<-data.frame(x=c(1:10),y=rnorm(10),z=c(10:1)),c("y","x"),c("x","y")))
(d<-rename.columns(d,c("y","x"),c("x","y")))
}
