% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/life.tools.R
\name{rename.list}
\alias{rename.list}
\title{RENAME LIST}
\usage{
rename.list(list, old.names, new.names)
}
\arguments{
\item{list}{named list}

\item{old.names}{old names}

\item{new.names}{new names}
}
\value{
dataframe with replaced column names
}
\description{
rename.columns( data, c( "alter", "groesse" ), c( "age", "height" ) ) takes a data frame d containing the old column name "alter" and "groesse" and returns a new dataframe with the replaced names "age" and "size".
}
\examples{
(l<-rename.list(list(x="x",y="y",z="z"),c("y","z","x"),c("Ypsilon","CED","U")))
}
