% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/life.tools.R
\name{sum.na}
\alias{sum.na}
\title{SUM MISSINGS}
\usage{
\method{sum}{na}(data)
}
\arguments{
\item{data}{dataframe for which missings in columns should be summed up}
}
\value{
dataframe with the number of missing data for each column
}
\description{
gives the number of the missing values of every column of a dataframe
}
\examples{
(d<-data.frame(x=c(NA,"Hello",NA,"World",NA),y=c(1:5),z=rep(NA,5)))
sum.na(d)
}
