% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/life.tools.R
\name{table.df}
\alias{table.df}
\title{TABLE DATAFRAME}
\usage{
table.df(data, summary = F, horizontal = T)
}
\arguments{
\item{data}{dataframe which columns should be summarised}

\item{summary}{logical: if is true additionally min, median, mean and max will be tabled}

\item{horizontal}{logical: result is shown in horizontal or vertical style}
}
\value{
table of data containing sum of missing and available data their mins and maxs.
}
\description{
table.df returns data about missings, availables of every column in a given
dataframe. If  summary is TRUE, min, max, median and mean are shown additionally.
}
\examples{
table.df(data.frame(x=c(1:3),y=c(NA,1,NA),z=c(NA,NA,NA)),F,F,F)
table.df(data.frame(x=c(1:3),y=c(NA,1,NA),z=c(NA,NA,NA)),T,F,F)
table.df(data.frame(x=c(1:3),y=c(NA,1,NA),z=c(NA,NA,NA)),F,T,F)
table.df(data.frame(x=c(1:3),y=c(NA,1,NA),z=c(NA,NA,NA)),T,T,F)
table.df(data.frame(x=c(1:3),y=c(NA,1,NA),z=c(NA,NA,NA)),F,F,T)
table.df(data.frame(x=c(1:3),y=c(NA,1,NA),z=c(NA,NA,NA)),T,F,T)
table.df(data.frame(x=c(1:3),y=c(NA,1,NA),z=c(NA,NA,NA)),F,T,T)
table.df(data.frame(x=c(1:3),y=c(NA,1,NA),z=c(NA,NA,NA)),T,T,T)
}
